+++ CHANGELOG.md
@@ -3,10 +3,11 @@
 <!-- <START NEW CHANGELOG ENTRY> -->
 ## 0.7.0 (June XX, 2024)
 * Converted to a Jupyter Lab 4.0+ compatible plugin.
+* Switched to Jupyter compatible BSD license.
 * Renamed to JPSL Instructor Tools (JPSLInstructorTools) to better reflect 
   the tight integration with Jupyter Physical Science Lab.
 * Added capability to enable and disable the Instructor Tools menu as well 
-  as show and hide it , so that this plugin can exist in a Jupyter Hub 
+  as show and hide it. This allows the plugin to exist in a Jupyter Hub 
   installation used by both the instructor and students. By default, the menu 
   is deactivated and will not activate if a student has a notebook set to 
   disallow use of instructor tools.